CU-86euem7cx_Seed-Data-in-database_Nasrin-Akter

+++ src/company/company.controller.ts
@@ -21,12 +21,14 @@ import { User } from 'src/auth/decorators/user.decorator';
 import { JwtAuthGuard } from 'src/auth/guards/jwt-auth.guard';
 import { VerifyOtpDto } from './dto/verify-otp.dto';
 import { OptionalJwtAuthGuard } from 'src/auth/guards/optional-jwt-auth.guard';
+import { Permissions } from '@/auth/decorators/permissions.decorator';
 
+@UseGuards(OptionalJwtAuthGuard)
 @Controller('companies')
 export class CompanyController {
   constructor(private readonly companyService: CompanyService) {}
 
-  @UseGuards(OptionalJwtAuthGuard)
+  @Permissions('company_create')
   @Post()
   create(
     @User('user') user: any,
@@ -41,45 +43,48 @@ export class CompanyController {
     return this.companyService.create(rest, adminInfo, sendOtpType);
   }
 
+  // @UseGuards(OptionalJwtAuthGuard)
+  @Permissions('company_list')
   @Get()
-  findAll(@Query() query: GenericQueryDto) {
+  findAll(@Query() query: GenericQueryDto, @User('user') user: any) {
     return this.companyService.findAll(query);
   }
 
+  // @UseGuards(OptionalJwtAuthGuard)
+  @Permissions('company_view')
   @Get(':id')
-  findOne(@Param('id') id: string) {
+  findOne(@Param('id') id: string, @User('user') user: any) {
     return this.companyService.findOne(+id);
   }
 
   @UseGuards(JwtAuthGuard)
+  @Permissions('company_approve')
   @Patch(':id/approve')
   approveCompany(@User('user') user: any, @Param('id') id: string) {
-    console.log(user.userId, 'User ID from JWT');
+    if (user.role.roleName !== 'SUPER_ADMIN') {
+      throw new UnauthorizedException('Only super admin can approve a company');
+    }
     return this.companyService.approveCompany(+id, user.userId);
   }
 
+  @UseGuards(JwtAuthGuard)
+  @Permissions('company_delete')
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  remove(@Param('id') id: string, @User('user') user: any) {
+    if (user.role.roleName !== 'SUPER_ADMIN') {
+      throw new UnauthorizedException('Only super admin can delete a company');
+    }
     return this.companyService.remove(+id);
   }
 
+  // @UseGuards(OptionalJwtAuthGuard)
+  @Permissions('company_update')
   @Put(':id')
   update(@Param('id') id: string, @Body() dto: UpdateCompanyDto) {
     return this.companyService.update(+id, dto);
   }
 
-  // Utility: Extracts and checks if the user is SUPER_ADMIN
-  private extractIsSuperAdmin(authHeader: string | undefined): boolean {
-    const token = authHeader?.split(' ')[1];
-    try {
-      const decoded: any = jwt.decode(token);
-      const roleName = decoded?.role?.roleName;
-      return roleName === 'SUPER_ADMIN' ? true : false;
-    } catch (error) {
-      throw new UnauthorizedException('Failed to decode token');
-    }
-  }
-
+  @Permissions('verify-otp')
   @Post('verify-otp')
   verifyOtp(@Body() body: VerifyOtpDto) {
     return this.companyService.verifyOtp(body);
